Move some templates to more appropriate locations.
