+++ src/main/kotlin/com/studyshare/templates/General.kt
@@ -1,18 +1,8 @@
 package com.studyshare.templates
 
 import com.studyshare.attachment.AttachmentView
-import com.studyshare.solution.Solution
-import com.studyshare.solution.VoteUpdate
-import com.studyshare.task.Task
-import com.studyshare.utils.Post
 import kotlinx.html.*
 import org.bson.types.ObjectId
-
-enum class AccessLevel {
-    NONE,
-    DELETE,
-    EDIT
-}
 
 fun FlowContent.nonImageAttachmentTemplate(nonImageAttachments: List<AttachmentView>) {
     if (nonImageAttachments.isNotEmpty()) {
@@ -63,57 +53,6 @@ fun FlowContent.fancyboxSetupScript() {
     }
 }
 
-fun FlowContent.votingButton(
-    isActive: Boolean,
-    voteType: String,
-    callbackId: ObjectId,
-    voteUrl: String,
-    icon: String
-) {
-    button {
-        classes = setOf("btn", "outline", "solution-voting")
-        classes +=
-            if (isActive) "primary"
-            else "secondary"
-
-        attributes["hx-post"] = voteUrl
-        attributes["hx-target"] = "#voting-${callbackId.toHexString()}"
-        attributes["hx-swap"] = "outerHTML"
-        span {
-            classes = setOf("material-symbols-rounded", voteType)
-            +icon
-        }
-    }
-}
-
-fun FlowContent.votingTemplate(voteUpdate: VoteUpdate, callbackId: ObjectId) {
-    div {
-        classes = setOf("solution-voting")
-        id = "voting-${callbackId.toHexString()}"
-
-        votingButton(
-            isActive = voteUpdate.isUpvoted,
-            voteType = "upvote",
-            callbackId = callbackId,
-            voteUrl = "/solutions/${callbackId.toHexString()}/upvote",
-            icon = "arrow_upward"
-        )
-
-        span {
-            classes = setOf("vote-count")
-            +voteUpdate.voteCount.toString()
-        }
-
-        votingButton(
-            isActive = voteUpdate.isDownvoted,
-            voteType = "downvote",
-            callbackId = callbackId,
-            voteUrl = "/solutions/${callbackId.toHexString()}/downvote",
-            icon = "arrow_downward"
-        )
-    }
-}
-
 fun FlowContent.contentLoadTemplate(url: String) {
     div {
         attributes["hx-get"] = url
@@ -140,25 +79,9 @@ fun FlowContent.deletionButton(getUrl: String) {
     }
 }
 
-fun FlowContent.postDeletionButton(post: Post) {
-    val url = when (post) {
-        is Solution -> "/solutions/deletion-modal?solutionId=${post.id}"
-        is Task -> "/${post.groupId}/deletion-modal?taskId=${post.id}"
-        else -> throw IllegalArgumentException("Invalid post")
-    }
-
-    deletionButton(url)
-}
-
-fun FlowContent.postEditingButton(post: Post) {
-    val url = when (post) {
-        is Task -> "/tasks/editing-modal?taskId=${post.id}"
-        is Solution -> "/solutions/editing-modal?id=${post.id}"
-        else -> throw IllegalArgumentException("Invalid post")
-    }
-
+fun FlowContent.editButton(getUrl: String) {
     button(classes = "btn secondary outline") {
-        attributes["hx-get"] = url
+        attributes["hx-get"] = getUrl
         attributes["hx-target"] = "body"
         attributes["hx-swap"] = "beforeend"
 
@@ -168,25 +91,15 @@ fun FlowContent.postEditingButton(post: Post) {
     }
 }
 
-fun FlowContent.postActions(post: Post, accessLevel: AccessLevel)
-{
-    when (accessLevel) {
-        AccessLevel.NONE -> return
-        AccessLevel.DELETE -> postDeletionButton(post)
-        AccessLevel.EDIT -> {postEditingButton(post); postDeletionButton(post)}
-    }
-}
-
-
 fun FlowContent.localDateSpan(objectId: ObjectId) {
     val script = """
         on load 1
-            put convertUTCDateToLocalDate(me.dataset.date) into me
+            put convertUTCDateToLocalDate(me.dataset.timestamp) into me
         end
     """.trimIndent()
     span {
         attributes["_"] = script
-        attributes["data-date"] = objectId.timestamp.toString()
+        attributes["data-timestamp"] = objectId.timestamp.toString()
     }
 }
