+++ src/main/kotlin/com/studyshare/templates/Solution.kt
@@ -5,6 +5,58 @@ import com.studyshare.solution.SolutionView
 import com.studyshare.solution.VoteUpdate
 import kotlinx.html.FlowContent
 import kotlinx.html.*
+import org.bson.types.ObjectId
+
+fun FlowContent.votingButton(
+    isActive: Boolean,
+    voteType: String,
+    callbackId: ObjectId,
+    voteUrl: String,
+    icon: String
+) {
+    button {
+        classes = setOf("btn", "outline", "solution-voting")
+        classes +=
+            if (isActive) "primary"
+            else "secondary"
+
+        attributes["hx-post"] = voteUrl
+        attributes["hx-target"] = "#voting-${callbackId.toHexString()}"
+        attributes["hx-swap"] = "outerHTML"
+        span {
+            classes = setOf("material-symbols-rounded", voteType)
+            +icon
+        }
+    }
+}
+
+fun FlowContent.votingTemplate(voteUpdate: VoteUpdate, callbackId: ObjectId) {
+    div {
+        classes = setOf("solution-voting")
+        id = "voting-${callbackId.toHexString()}"
+
+        votingButton(
+            isActive = voteUpdate.isUpvoted,
+            voteType = "upvote",
+            callbackId = callbackId,
+            voteUrl = "/solutions/${callbackId.toHexString()}/upvote",
+            icon = "arrow_upward"
+        )
+
+        span {
+            classes = setOf("vote-count")
+            +voteUpdate.voteCount.toString()
+        }
+
+        votingButton(
+            isActive = voteUpdate.isDownvoted,
+            voteType = "downvote",
+            callbackId = callbackId,
+            voteUrl = "/solutions/${callbackId.toHexString()}/downvote",
+            icon = "arrow_downward"
+        )
+    }
+}
 
 fun FlowContent.solutionTemplate(solutionView: SolutionView, accessLevel: AccessLevel) {
     val images = solutionView.attachments.filter { attachmentView: AttachmentView -> attachmentView.attachment.isImage }
